pyproject.toml
[project]
name = "p7-g2-multiclase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "evidently>=0.4.36",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "optuna>=4.8.0",
    "pandas>=2.2,<3",
    "pillow>=10,<11",
    "plotly>=6.6.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.55.0",
    "supabase>=2.28.3",
    "tensorflow==2.21.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "xgboost>=3.2.0",
]

[dependency-groups]
dev = [
    "pytest-cov>=7.1.0",
]
